Schema Fingerprints/Timestamps; Oracle NUMBER vs DECIMAL, Savepoints timestamp
Changed Savepoints write to save the last timestamp actually written any time JDBC is loaded. If ONLY file is loaded, it works as it did OOTB.

Changed fingerprint/timestamp determination of what combination needs to be loaded on subsequent load. Look for additional timestamps in the fingerprint we start in, and if none there move to the next fingerprint.

ORACLE - The initial insert puts integers as NUMBER() columns, but the schema evaluation with the jdbc oracle driver produces DECIMAL() , causing schema compare to always fail. Forced to NUMBER() so evaluation happens properly.

Added new JDBCOutputWriter.scala class. It is an empty shell - only commented code at this point. Next commit/push will contain changes (in theory).