Schema handling for PostgreSQL; DATETIME handling for SQL Server
PostgreSQL had previously only been tested with separate databases, each using the "public" schema. In testing with one database with multiple schemas there were issues because the dialect was used to grab the schema and it always returned "public" rather than the actual schema to be used. The schema is stored in the config.yaml file for each database connection, and that value was leveraged instead of the default value for the dialect.

SQL Server JDBC driver data type conversion of "TimestampType" by default maps to DATETIME, which with bad values produces an out of range error. Example was a year of "0001" which will not map properly. Changed to force DATETIME2 for "TimestampType" to avoid out of range errors. PostgreSQL handles it fine.